+++ script_ejemplo_aplicacion.py
@@ -1,0 +1,7 @@
+from menu import Menu
+
+def main():
+    menu = Menu()
+    menu.iniciar_sistema()
+
+main()
